Fix extensive test failures related to Integer Precision and FinanceSystem mocking.
- **FinanceSystem Integration**: Added `outstanding_bonds` property to `FinanceSystem` for legacy compatibility. Updated `tests/conftest.py` to safely mock deprecated methods.
- **Integer Precision Enforcement**:
    - Updated `Bootstrapper` (simulation/systems/bootstrapper.py) to use integers for capital injection and wages.
    - Updated `HouseholdFactory` (simulation/factories/agent_factory.py) to enforce integer types for initial assets.
    - Updated `InfrastructureManager` (modules/government/components/infrastructure_manager.py) to cast costs to `int`.
    - Updated ALL Transaction Handlers (`goods`, `labor`, `financial`, `stock`) in `simulation/systems/handlers/` to calculate trade values as `int` before calling `SettlementSystem`.
- **Household State Fixes**:
    - Updated `AssetManager` (simulation/decisions/household/asset_manager.py) to handle `current_wage` vs `current_wage_pennies` deprecation.
    - Fixed `test_consumption_manager.py` to usage integer pennies for wallet and wages.
- **Test Fixes**:
    - Updated `tests/system/test_phase29_depression.py` to pass integer to `deposit`.
    - Verified `test_agent_decision.py` and `test_atomic_settlement.py` pass.

+++ simulation/systems/handlers/financial_handler.py
@@ -17,7 +17,7 @@ class FinancialTransactionHandler(ITransactionHandler):
 
     def handle(self, tx: Transaction, buyer: Any, seller: Any, context: TransactionContext) -> bool:
         tx_type = tx.transaction_type
-        trade_value = tx.quantity * tx.price
+        trade_value = int(tx.quantity * tx.price)
 
         success = False
 
@@ -25,13 +25,14 @@ class FinancialTransactionHandler(ITransactionHandler):
              success = context.settlement_system.transfer(buyer, seller, trade_value, tx_type)
 
              if success and isinstance(buyer, Firm):
-                 buyer.record_expense(trade_value, tx.currency)
+                 buyer.record_expense(int(trade_value), tx.currency)
 
         elif tx_type == "dividend":
              success = context.settlement_system.transfer(seller, buyer, trade_value, "dividend_payment")
 
              if success and isinstance(buyer, Household) and hasattr(buyer, "capital_income_this_tick"):
-                 buyer.capital_income_this_tick += trade_value
+                 # capital_income_this_tick is int in EconStateDTO
+                 buyer.capital_income_this_tick += int(trade_value)
 
         elif tx_type == "tax":
             # Atomic Settlement to Government
@@ -40,14 +41,14 @@ class FinancialTransactionHandler(ITransactionHandler):
             # TransactionProcessor logic: settle_atomic(buyer, [(gov, amount, item_id)])
 
             gov = context.government
-            credits = [(gov, trade_value, tx.item_id)]
+            credits = [(gov, int(trade_value), tx.item_id)]
 
             success = context.settlement_system.settle_atomic(buyer, credits, context.time)
 
             if success:
                  gov.record_revenue({
                          "success": True,
-                         "amount_collected": trade_value,
+                         "amount_collected": int(trade_value),
                          "tax_type": tx.item_id,
                          "payer_id": buyer.id,
                          "payee_id": gov.id,
@@ -56,6 +57,6 @@ class FinancialTransactionHandler(ITransactionHandler):
 
                  # WO-116 Fix: Ensure Firms record tax as expense for accounting integrity
                  if isinstance(buyer, Firm):
-                     buyer.record_expense(trade_value, tx.currency)
+                     buyer.record_expense(int(trade_value), tx.currency)
 
         return success is not None
